requirements.txt
streamlit>=1.36.0
pillow>=10.3.0
opencv-python>=4.9.0
pytesseract>=0.3.10
numpy>=1.26.0
